Refatoração da main e criacao de novo arquivo de cadastro para melhorar modularidade do código
